Most Disks refactoring should be done. Still some errors.

+++ yew-app/src/ai/ai.rs
@@ -1,5 +1,5 @@
 use crate::util::util::{DiskColor, Disks};
 
 pub trait AI {
-    fn get_move(&self, board: &Disks, player: DiskColor) -> usize;
+    fn get_move(&self, board: &Disks, player: DiskColor) -> u8;
 }
